kup: Skip missing directories

+++ build.bash
@@ -2,6 +2,8 @@
 
 [[ -f localconfig.bash ]] && source localconfig.bash
 
+shopt -s nullglob
+
 die() {
 	echo "${@}" >&2
 	tput bel
@@ -421,23 +423,25 @@ do_rsync() {
 do_kup() {
 	local arch pkg new_fn major_ver dir
 	for arch in amd64 arm64 ppc64le x86; do
-		cd ~/binpkg/"${arch}"/kernel/sys-kernel/gentoo-kernel ||
-			die "cd failed"
-		for pkg in *.gpkg.tar; do
-			new_fn=${pkg/.gpkg/.${arch}.gpkg}
-			major_ver=${pkg%.*-*}
-			major_ver=${major_ver#gentoo-kernel-}
-			dir=/pub/proj/dist-kernel/binpkg/${arch}/${major_ver}
+		if [[ -d ~/binpkg/${arch}/kernel/sys-kernel/gentoo-kernel ]]; then
+			cd ~/binpkg/"${arch}"/kernel/sys-kernel/gentoo-kernel ||
+				die "cd failed"
+			for pkg in *.gpkg.tar; do
+				new_fn=${pkg/.gpkg/.${arch}.gpkg}
+				major_ver=${pkg%.*-*}
+				major_ver=${major_ver#gentoo-kernel-}
+				dir=/pub/proj/dist-kernel/binpkg/${arch}/${major_ver}
 
-			# NB: kup's exit status is always zero (read: useless)
-			if ! grep "${new_fn}$" <(kup ls "${dir}"); then
-				ln "${pkg}" "${new_fn}" || die "ln failed"
-				gpg --detach-sign "${new_fn}" || die "gpg failed"
-				kup mkdir "${dir}" || die "kup mkdir failed"
-				kup putraw "${new_fn}" "${new_fn}.sig" "${dir}/" || die "kup putraw failed"
-				rm "${new_fn}" "${new_fn}.sig" || die "rm temp failed"
-			fi
-		done
+				# NB: kup's exit status is always zero (read: useless)
+				if ! grep "${new_fn}$" <(kup ls "${dir}"); then
+					ln "${pkg}" "${new_fn}" || die "ln failed"
+					gpg --detach-sign "${new_fn}" || die "gpg failed"
+					kup mkdir "${dir}" || die "kup mkdir failed"
+					kup putraw "${new_fn}" "${new_fn}.sig" "${dir}/" || die "kup putraw failed"
+					rm "${new_fn}" "${new_fn}.sig" || die "rm temp failed"
+				fi
+			done
+		fi
 	done
 }
